# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/feed": {
            "get": {
                "tags": [
                    "基础接口"
                ],
                "summary": "视频流接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "latest_time",
                        "name": "latest_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.FeedVideoRespStruct"
                        }
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "tags": [
                    "公共接口"
                ],
                "summary": "首页",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publish/action/": {
            "post": {
                "tags": [
                    "基础接口"
                ],
                "summary": "视频投稿",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "data",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.UploadRespStruct"
                        }
                    }
                }
            }
        },
        "/publish/list/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础接口"
                ],
                "summary": "发布列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录用户的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查找目标用户的id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status_msg为成功",
                        "schema": {
                            "$ref": "#/definitions/controller.GetPublishListResponse"
                        }
                    }
                }
            }
        },
        "/relation/action/": {
            "post": {
                "tags": [
                    "社交接口"
                ],
                "summary": "关注操作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "to_user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "关注操作",
                        "name": "action_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.FollowRespStruct"
                        }
                    }
                }
            }
        },
        "/relation/follow/list/": {
            "get": {
                "tags": [
                    "社交接口"
                ],
                "summary": "关注列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.FollowListRespStruct"
                        }
                    }
                }
            }
        },
        "/relation/follower/list/": {
            "get": {
                "tags": [
                    "社交接口"
                ],
                "summary": "粉丝列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.FollowerListRespStruct"
                        }
                    }
                }
            }
        },
        "/relation/friend/list/": {
            "get": {
                "tags": [
                    "社交接口"
                ],
                "summary": "好友列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.FirendListRespStruct"
                        }
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "tags": [
                    "基础接口"
                ],
                "summary": "用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.UserInfoRespStruct"
                        }
                    }
                }
            }
        },
        "/user/login/": {
            "post": {
                "tags": [
                    "基础接口"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.UserLoginRespStruct"
                        }
                    }
                }
            }
        },
        "/user/register/": {
            "post": {
                "tags": [
                    "基础接口"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.FeedVideoRespStruct": {
            "type": "object",
            "properties": {
                "next_time": {
                    "type": "integer"
                },
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.VideoInfo"
                    }
                }
            }
        },
        "controller.FirendListRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "user_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.User"
                    }
                }
            }
        },
        "controller.FollowListRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "user_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.User"
                    }
                }
            }
        },
        "controller.FollowRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "controller.FollowerListRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "user_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.User"
                    }
                }
            }
        },
        "controller.GetPublishListResponse": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.VideoInfo"
                    }
                }
            }
        },
        "controller.UploadRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "controller.UserInfoRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/service.Author"
                }
            }
        },
        "controller.UserLoginRespStruct": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "service.Author": {
            "type": "object",
            "properties": {
                "follow_count": {
                    "type": "integer"
                },
                "follower_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_follow": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.User": {
            "type": "object",
            "properties": {
                "follow_count": {
                    "type": "integer"
                },
                "follower_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_follow": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.VideoInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/service.Author"
                        }
                    ]
                },
                "comment_count": {
                    "description": "评论数",
                    "type": "integer"
                },
                "cover_url": {
                    "description": "封面路径",
                    "type": "string"
                },
                "favorite_count": {
                    "description": "点赞数",
                    "type": "integer"
                },
                "id": {
                    "description": "视频唯一标识",
                    "type": "integer"
                },
                "is_favorite": {
                    "description": "是否点赞",
                    "type": "boolean"
                },
                "play_url": {
                    "description": "视频路径",
                    "type": "string"
                },
                "title": {
                    "description": "视频标题",
                    "type": "string"
                }
            }
        }
    }
}